requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.9.0

# Donor reward stack (Phase 2 + 3a)
solders>=0.21,<0.30
solana>=0.34
xrpl-py>=4.0
